+++ src/main/kotlin/com/github/sleepypanda/feesh/events/models/ArmorStandCustomNameChangedEvent.kt
@@ -1,13 +1,23 @@
 package com.github.sleepypanda.feesh.events.models
 
-import net.minecraft.world.entity.decoration.ArmorStand
-
 /**
  * Fired when an ArmorStand's custom name is updated (e.g. became non-empty after being loaded, or mob HP is changed).
  */
 data class ArmorStandCustomNameChangedEvent(
-    val entity: ArmorStand,
     val entityId: Int,
-    val customNameFormatted: String,
-    val customNameUnformatted: String
-)
+    val isFirstLoaded: Boolean, // if first saw the armor stand, its previous custom name will be empty
+    val previousCustomName: CustomName,
+    val customName: CustomName,
+    val position: Position,
+) {
+    data class CustomName(
+        val formatted: String,
+        val unformatted: String,
+    )
+
+    data class Position(
+        val x: Double,
+        val y: Double,
+        val z: Double,
+    )
+}