+++ src/main/kotlin/com/github/sleepypanda/feesh/features/rendering/RareMobHighlight.kt
@@ -3,7 +3,7 @@ package com.github.sleepypanda.feesh.features.rendering
 import com.github.sleepypanda.feesh.FeeshMod
 import com.github.sleepypanda.feesh.constants.SeaCreatures
 import com.github.sleepypanda.feesh.events.EventBus
-import com.github.sleepypanda.feesh.events.models.ArmorStandDetailsLoadedEvent
+import com.github.sleepypanda.feesh.events.models.ArmorStandCustomNameChangedEvent
 import com.github.sleepypanda.feesh.events.models.ClientTickEvent
 import com.github.sleepypanda.feesh.events.models.WorldChangedEvent
 import com.github.sleepypanda.feesh.settings.categories.WorldRendering
@@ -34,7 +34,7 @@ object RareMobHighlight {
     fun init() {
         EventBus.subscribe(WorldChangedEvent::class, ::onWorldChange)
         EventBus.subscribe(ClientTickEvent::class, ::onClientTick)
-        EventBus.subscribe(ArmorStandDetailsLoadedEvent::class, ::onArmorStandDetailsLoaded)
+        EventBus.subscribe(ArmorStandCustomNameChangedEvent::class, ::onArmorStandCustomNameChanged)
         updateEnabledMobTypes()
     }
 
@@ -54,10 +54,20 @@ object RareMobHighlight {
         enabledMobTypes = WorldRendering.highlightSeaCreaturesList.map { it.displayName }.distinct().toList()
     }
 
-    private fun onArmorStandDetailsLoaded(event: ArmorStandDetailsLoadedEvent) {
+    private fun onArmorStandCustomNameChanged(event: ArmorStandCustomNameChangedEvent) {
+        if (!event.isFirstLoaded) return
         if (!WorldRendering.highlightSeaCreatures || !WorldUtils.isInSkyblock() || !WorldUtils.isInFishingWorld()) return
-        val entity = event.entity
-        val cleanName = EntityUtils.parseSeaCreatureNametag(entity, enabledMobTypes)?.baseMobName ?: return
+
+        val world = FeeshMod.mc.level ?: return
+        val cleanName = EntityUtils.parseSeaCreatureNametag(
+            entityId = event.entityId,
+            customNameFormatted = event.customName.formatted,
+            customNameUnformatted = event.customName.unformatted,
+            x = event.position.x,
+            y = event.position.y,
+            z = event.position.z,
+            includedSeaCreatureNames = enabledMobTypes,
+        )?.baseMobName ?: return
         if (!enabledMobTypes.contains(cleanName)) return
 
         val scInfo = SeaCreatures.allSeaCreatures.find { it.name == cleanName }
@@ -75,11 +85,11 @@ object RareMobHighlight {
         val entities: MutableList<Entity> = mutableListOf()
 
         // Volcanic Snail and Jumpin Jack are ItemDisplay entities instead of LivingEntity
-        val potentialMobEntity = entity.level().getEntity(entity.id - mobEntityShift)
+        val potentialMobEntity = world.getEntity(event.entityId - mobEntityShift)
         var mobEntity = if (potentialMobEntity is LivingEntity || potentialMobEntity is ItemDisplay) potentialMobEntity else return
 
         if (cleanName == HighlightableSeaCreatureTypes.JAWBUS_FOLLOWER.displayName && mobEntity is Slime && mobEntity !is MagmaCube) { // Fire Eel
-            mobEntity = entity.level().getEntity(entity.id - 11) as? LivingEntity ?: return // -1 is for tail, we want to find Fire Eel's head
+            mobEntity = world.getEntity(event.entityId - 11) as? LivingEntity ?: return // -1 is for tail, we want to find Fire Eel's head
         }
 
         if (mobEntity is LivingEntity && !mobEntity.isAlive) return
@@ -117,7 +127,7 @@ object RareMobHighlight {
         if (cleanName == HighlightableSeaCreatureTypes.WIKI_TIKI.displayName) {
             val wikiTikiEntitiesShifts = listOf(3, 5, 7)
             wikiTikiEntitiesShifts.forEach { shift ->
-                val prevEntity = entity.level().getEntity(entity.id - shift) as? LivingEntity ?: return@forEach
+                val prevEntity = world.getEntity(event.entityId - shift) as? LivingEntity ?: return@forEach
                 entities.add(prevEntity)
             }
         }